- Renaming of foreign key database columns - Server url changes to match with context path - New endpoint to create addresses for current logged-in user

+++ backend/src/main/resources/db/migration/V20260524_2125__createShoppingCartTables.sql
@@ -7,7 +7,7 @@ create table shopping_carts
     version bigint not null default 0,
     shopping_carts_id serial primary key,
     price numeric(10, 2) not null,
-    user_id integer not null,
+    user_id bigint not null,
     constraint fk_user
         foreign key (user_id) references users (users_id)
 );
@@ -21,8 +21,8 @@ create table cart_items
     version bigint not null default 0,
     cart_items_id serial primary key,
     quantity integer not null,
-    shopping_cart_id integer not null,
-    product_id integer not null,
+    shopping_cart_id bigint not null,
+    product_id bigint not null,
     constraint fk_shopping_cart
         foreign key (shopping_cart_id) references shopping_carts (shopping_carts_id),
     constraint fk_product
